requirements.txt
fastapi
uvicorn
tensorflow  # Use the CPU version to avoid CUDA issues
numpy
pillow
pyngrok
requests
python-multipart
torch
ultralytics 